+++ src/components/HypothesisTesting/HTSimulation.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useState }  from "react";
 import PerformTest from "./PerformTest.js";
-import { Alert, Container, Row, Col, Button, Input, InputGroup, InputGroupAddon, InputGroupText } from "reactstrap";
+import { Alert, Container, Row, Button } from "reactstrap";
 import TestInputs from "./TestInputs.js";
-import HypothesisDropdown from "./HypothesisDropdown.js";
+import HypothesisSelector from "./HypothesisSelector.js";
 import _ from "lodash";
 
 export default function HTSimulation() {
@@ -28,7 +28,7 @@ export default function HTSimulation() {
               <p>The true population distribution will be revealed at the end.</p>
               {(testType === "oneSample") ? (
                 <p>
-                  Suppose that our farmer has changed the variety of feed the cows eat. It might be reasonable to think that the cows now produce more or less milk than they had before. As a researcher, what assertion would you like to make about these cows’ milk production now? Choose an Option and specify a hypothesized amount. To help make an informed guess, note that the distribution of millk production before we changed the feed had a mean of about 64 gallons
+                  Suppose that our farmer has changed the variety of feed the cows eat. It might be reasonable to think that the cows now produce more or less milk than they had before. As a researcher, what assertion would you like to make about these cows’ milk production now? Choose an Option and specify a hypothesized amount. To help make an informed guess, note that the distribution of milk production before we changed the feed had a mean of about 64 gallons
                 </p>
               ) : (
                 <p>
@@ -38,28 +38,8 @@ export default function HTSimulation() {
             </Alert>
           </Row>
           <br/>
-          <Row>
-            <Col xs="auto">
-              <HypothesisDropdown testType={testType} setHypothesis={setHypothesis}/>
-            </Col>
-            <Col>
-              {(testType === "oneSample") && (
-                <InputGroup>
-                  <Input
-                    className="Center"
-                    type="number"
-                    value={mue0}
-                    step={1}
-                    min={1}
-                    max={1000}
-                    onChange={(event) => setMue0(event.target.value)}
-                  />
-                  <InputGroupAddon addonType="append">
-                    <InputGroupText>gallons of milk per day.</InputGroupText>
-                  </InputGroupAddon>
-                </InputGroup>
-              )}
-            </Col>
+          <Row style={{width: "80%", margin: "auto"}}>
+            <HypothesisSelector testType={testType} setHypothesis={setHypothesis} mue0={mue0} setMue0={setMue0}/>
           </Row>
           <br/>
           <Button color="primary" onClick={() => setStage(3)}> Continue </Button>
